requirements.txt
openai
python-dotenv
requests
pypdf
tavily-python